Abstracted common summary module behavior to a base class. Refactored scientificsummarymodule.
