+++ test/test.js
@@ -41,6 +41,7 @@ require('./spec/core/DownloadViewTest');
 require('./spec/core/EventPageTest');
 require('./spec/core/FormatterTest');
 require('./spec/core/LinkProductViewTest');
+require('./spec/core/SummaryModuleTest');
 require('./spec/core/TextProductViewTest');
 
 require('./spec/dyfi/DYFIFormModuleTest');
